Main v3

+++ Point_of_Sale/Point_of_Sale/Repository/PointOfSaleRepository.cs
@@ -7,6 +7,10 @@ using Point_of_Sale.Interface;
 using Point_of_Sale.Models;
 using Point_of_Sale.Models.DBContext;
 using System.Globalization;
+using AspNetCore.ReportingServices.ReportProcessing.ReportObjectModel;
+using static iTextSharp.text.pdf.AcroFields;
+using Point_of_Sale.DTO;
+using System.Web.Helpers;
 
 namespace Point_of_Sale.Repository
 {
@@ -14,11 +18,13 @@ namespace Point_of_Sale.Repository
     {
         private readonly PointOfSaleDbContext db;
         private readonly IWebHostEnvironment _hostingEnvironment;
+        private readonly IGlobal global;
 
-        public PointOfSaleRepository(PointOfSaleDbContext context, IWebHostEnvironment hostingEnvironment)
+        public PointOfSaleRepository(PointOfSaleDbContext context, IWebHostEnvironment hostingEnvironment, IGlobal global_rep)
         {
             db = context;
             _hostingEnvironment = hostingEnvironment;
+            global = global_rep;
         }
 
         public int SaveInvoice(int UserId)
@@ -74,30 +80,9 @@ namespace Point_of_Sale.Repository
                     db.tbl_sales.Add(sales);
                     db.SaveChanges();
                 }
-
-                db.tbl_cart.RemoveRange(cart);
-                db.SaveChanges();
-
-                return true;
-            }
-            catch (Exception)
-            {
-                return false;
-            }
-        }
-
-        public bool DeductItem(int InvoiceId)
-        {
-            var cart = db.tbl_invoice.Where(x => x.Id == InvoiceId).SingleOrDefault();
-            try
-            {
-                var sales = db.tbl_sales.Where(x => x.InvoiceId == InvoiceId).ToList();
-                foreach (var item in sales)
+                if (cart != null)
                 {
-                    var totalQty = GetAvailableItem(item.Quantity, item.ProductId);
-
-                    var product = db.tbl_item.Where(x => x.Id == item.ProductId).SingleOrDefault();
-                    product.Quantity = totalQty;
+                    db.tbl_cart.RemoveRange(cart);
                     db.SaveChanges();
                 }
                 return true;
@@ -108,54 +93,57 @@ namespace Point_of_Sale.Repository
             }
         }
 
+        public ResultDTO DeductQuantity(int InvoiceId)
+        {
+            ResultDTO result = new ResultDTO();
+            result.IsSuccess = true;
+            result.Message = "Success";
+
+            var sales = db.tbl_sales.Where(x => x.InvoiceId == InvoiceId).ToList();
+
+            foreach (var item in sales)
+            {
+                var dtls = db.tbl_item.Where(x => x.Id == item.ProductId && x.Quantity >= item.Quantity).FirstOrDefault();
+                if (dtls != null)
+                {
+                    dtls.Quantity = dtls.Quantity - item.Quantity;
+                    db.SaveChanges();
+                }
+                else
+                {
+                    var itemDesc = db.tbl_item.Where(x => x.Id == item.ProductId).FirstOrDefault().Description;
+                    if (itemDesc != null)
+                    {
+                        result.IsSuccess = false;
+                        result.Message = "The available quantity of " + itemDesc + " is not enough!";
+                        return result;
+                    }
+                }
+            }
+            return result;
+        }
         public decimal GetTotalAmount(int InvoiceId)
         {
             var TotalAmount = 0.00M;
             var invoice = db.tbl_invoice.Where(x => x.Id == InvoiceId).SingleOrDefault();
-            TotalAmount = invoice.AmountTotal;
-
+            if (invoice != null)
+            {
+                TotalAmount = invoice.AmountTotal;
+            }
             return TotalAmount;
         }
 
         public int GetAvailableItem(int Quantity, int ProductId)
         {
-            var TotalAmount = 0;
-            var product = db.tbl_item.Where(x => x.Id == ProductId).SingleOrDefault();
-            TotalAmount = product.Quantity - Quantity;
-
-            return TotalAmount;
+            var TotalQuantity = 0;
+            var product = global.GetItemDetails().Where(x => x.ProductId == ProductId).SingleOrDefault();
+            if (product != null)
+            {
+                TotalQuantity = product.Quantity - Quantity;
+            }
+            return TotalQuantity;
         }
 
-        public FileStreamResult GenerateReport()
-        {
-            // Create a new PDF document
-            iTextSharp.text.Document document = new iTextSharp.text.Document();
 
-            // Set the output file path
-            string relativePath = "reports/report.pdf";
-            relativePath = relativePath.Replace("\\", "/");
-            string outputPath = Path.Combine(_hostingEnvironment.WebRootPath, relativePath);
-
-            // Create a new PDF writer
-            PdfWriter writer = PdfWriter.GetInstance(document, new FileStream(outputPath, FileMode.Create));
-
-            // Open the PDF document
-            document.Open();
-
-            // Add content to the PDF document
-            document.Add(new Paragraph("Hello, World!"));
-
-            // Close the PDF document
-            document.Close();
-
-            MemoryStream stream = new MemoryStream();
-            byte[] bytes = stream.ToArray();
-            stream.Write(bytes, 0, bytes.Length);
-            stream.Position = 0;
-
-            //return outputPath;
-
-            return new FileStreamResult(stream,"application/pdf");
-        }
     }
 }
